Moved separate kind of endpoints to separate files (blueprints).
